+++ Editor/Discover/DiscoverWindow.cs
@@ -85,7 +85,7 @@ namespace GameplayIngredients.Editor
         {
             if (discoverAsset != null)
             {
-                var window = GetWindow<DiscoverWindow>(true);
+                var window = GetWindow<DiscoverWindow>(!discoverAsset.dockable);
                 window.SetDiscoverAsset(discoverAsset);
             }
             else
@@ -159,6 +159,11 @@ namespace GameplayIngredients.Editor
         /// <returns>true to discard, false to keep</returns>
         static bool FilterDiscoverObject(DiscoverAsset asset, Discover discoverObject)
         {
+            if (string.IsNullOrEmpty(asset.Tags))
+                asset.Tags = string.Empty;
+            if (string.IsNullOrEmpty(discoverObject.Tags))
+                discoverObject.Tags = string.Empty;
+
             string[] assetTags = asset.Tags.Split(' ');
             string[] objectTags = discoverObject.Tags.Split(' ');
 
@@ -345,6 +350,11 @@ namespace GameplayIngredients.Editor
                 GUILayout.Label(discoverAsset.Title, Styles.header);
                 using (new GUILayout.VerticalScope(Styles.indent))
                 {
+                    if(discoverAsset.Image != null)
+                    {
+                        DiscoverEditor.DrawImage(discoverAsset.Image);
+                    }
+
                     GUILayout.Label(discoverAsset.Description, Styles.body);
 
                     if(discoverAsset.Scenes != null)
@@ -353,6 +363,11 @@ namespace GameplayIngredients.Editor
                         {
                             using (new GroupLabelScope(map.Title))
                             {
+                                if(map.Image != null)
+                                {
+                                    DiscoverEditor.DrawImage(map.Image);
+                                }
+
                                 GUILayout.Label(map.Description, Styles.body);
 
                                 using (new GUILayout.HorizontalScope())
@@ -580,6 +595,8 @@ namespace GameplayIngredients.Editor
 
             public static GUIStyle tabContainer;
 
+            public static GUIStyle image;
+
             static Styles()
             {
                 header = new GUIStyle(EditorStyles.wordWrappedLabel);
@@ -605,8 +622,8 @@ namespace GameplayIngredients.Editor
 
                 boxHeader = new GUIStyle(GUI.skin.box);
                 boxHeader.normal.textColor = GUI.skin.label.normal.textColor;
-                boxHeader.fixedHeight = 20;
-                boxHeader.fontSize = 11;
+                boxHeader.fixedHeight = 24;
+                boxHeader.fontSize = 16;
                 boxHeader.fontStyle = FontStyle.Bold;
                 boxHeader.alignment = TextAnchor.UpperLeft;
                 boxHeader.margin = new RectOffset(0, 0, 0, 6);
@@ -623,6 +640,9 @@ namespace GameplayIngredients.Editor
 
                 tabContainer = new GUIStyle(EditorStyles.miniButton);
                 tabContainer.padding = new RectOffset(4, 4, 0, 0);
+
+                image = new GUIStyle(GUIStyle.none);
+                image.stretchWidth = true ;
 
             }
         }