+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 
 - Discover : Tag Filtering System
+- Discover : Images in Discover Window
 
 ## 2019.3.4